Hooking Up New Game
Refactoring map generation script (I kept the old one too, but labeled it as a standalone script) to accept the size of the map from the New Game menu.  Moved some of the code around to separate game rendering and stuff from the base map gen logic.  Basic coloring changes to the map rendering, too.